+++ test/my_setup.c
@@ -51,7 +51,7 @@ DECLARE_TEST(t_bug66548)
   sprintf((char*)attrs, "DSN=bug66548dsn;SERVER=%s;USER=%s;PASSWORD=%s;"
     "PORT=%d;DATABASE=%s;\0", myserver, myuid, mypwd, myport, mydb);
 
-  len= strlen(attrs);
+  len = (int)strlen(attrs);
 
   /* replacing ';' by '\0' */
   for (i= 0; i < len; ++i)
@@ -63,7 +63,7 @@ DECLARE_TEST(t_bug66548)
   /* Adding the extra string termination to get \0\0 */
   attrs[i]= '\0';
 
-  len= strlen(mydriver);
+  len = (int)strlen(mydriver);
 
   if (mydriver[0] == '{')
   {
